+++ framework/domain/src/main/java/is/codion/framework/domain/entity/Entity.java
@@ -353,6 +353,13 @@ public interface Entity extends Comparable<Entity> {
 		Builder clearPrimaryKey();
 
 		/**
+		 * Creates a new {@link Entity.Key.Builder} instance, initialized with
+		 * any primary key values found in this builder.
+		 * @return a new {@link Key.Builder}
+		 */
+		Key.Builder key();
+
+		/**
 		 * Builds the Entity instance
 		 * @return a new Entity instance
 		 */