+++ framework/domain/src/test/java/is/codion/framework/domain/entity/DefaultKeyTest.java
@@ -61,7 +61,7 @@ public final class DefaultKeyTest {
 
 	@Test
 	void singleKeyNull() {
-		Entity.Key key = ENTITIES.keyBuilder(Detail.TYPE).build();
+		Entity.Key key = ENTITIES.builder(Detail.TYPE).key().build();
 		assertTrue(key.isNull());
 		key = key.copy()
 						.with(Detail.ID, null)
@@ -86,7 +86,7 @@ public final class DefaultKeyTest {
 		Entity.Key deptKey = ENTITIES.primaryKey(Department.TYPE, 1);
 		assertNotEquals(empKey1, deptKey);
 
-		Entity.Key compMasterKey = ENTITIES.keyBuilder(CompositeMaster.TYPE)
+		Entity.Key compMasterKey = ENTITIES.builder(CompositeMaster.TYPE).key()
 						.with(CompositeMaster.COMPOSITE_MASTER_ID, 1)
 						.with(CompositeMaster.COMPOSITE_MASTER_ID_2, 2)
 						.build();
@@ -94,7 +94,7 @@ public final class DefaultKeyTest {
 		assertNotEquals(empKey1, compMasterKey);
 		assertNotEquals(new Object(), compMasterKey);
 
-		Entity.Key compMasterKey2 = ENTITIES.keyBuilder(CompositeMaster.TYPE)
+		Entity.Key compMasterKey2 = ENTITIES.builder(CompositeMaster.TYPE).key()
 						.with(CompositeMaster.COMPOSITE_MASTER_ID, 1)
 						.build();
 		assertNotEquals(compMasterKey, compMasterKey2);
@@ -162,7 +162,7 @@ public final class DefaultKeyTest {
 
 	@Test
 	void nullKeyEquals() {
-		Entity.Key nullKey = ENTITIES.keyBuilder(Employee.TYPE).build();
+		Entity.Key nullKey = ENTITIES.builder(Employee.TYPE).key().build();
 		Entity.Key zeroKey = ENTITIES.primaryKey(Employee.TYPE, 0);
 		assertNotEquals(nullKey, zeroKey);
 	}